feat(utils): infra change
Ref: #1040
