+++ packages/infra/lib/ihe-gateway-v2-stack.ts
@@ -10,7 +10,10 @@ import { createLambda } from "./shared/lambda";
 import { LambdaLayers } from "./shared/lambda-layers";
 import { Secrets } from "./shared/secrets";
 import { provideAccessToQueue } from "./shared/sqs";
-import { IQueue } from "aws-cdk-lib/aws-sqs";
+import { Queue } from "aws-cdk-lib/aws-sqs";
+import { SqsEventSource } from "aws-cdk-lib/aws-lambda-event-sources";
+import { SnsAction } from "aws-cdk-lib/aws-cloudwatch-actions";
+import { createQueue } from "./shared/sqs";
 
 interface IHEGatewayV2LambdasNestedStackProps extends NestedStackProps {
   lambdaLayers: LambdaLayers;
@@ -28,15 +31,71 @@ interface IHEGatewayV2LambdasNestedStackProps extends NestedStackProps {
   sentryDsn: string | undefined;
   iheResponsesBucketName: string;
   iheParsedResponsesBucketName: string;
-  writeToS3Lambda: Lambda;
-  writeToS3Queue: IQueue;
+  alarmAction?: SnsAction;
 }
+
+function settings() {
+  const writeToS3LambdaTimeout = Duration.seconds(25);
+  const writeToS3: QueueAndLambdaSettings = {
+    name: "IHEGatewayV2OutboundPatientDiscoveryWriteToS3",
+    entry: "ihe-outbound-patient-discovery-write-to-s3",
+    lambda: {
+      memory: 2048,
+      batchSize: 10,
+      timeout: writeToS3LambdaTimeout,
+      reportBatchItemFailures: true,
+    },
+    queue: {
+      alarmMaxAgeOfOldestMessage: Duration.hours(2),
+      maxMessageCountAlarmThreshold: 5_000,
+      maxReceiveCount: 3,
+      visibilityTimeout: Duration.seconds(writeToS3LambdaTimeout.toSeconds() * 2 + 1),
+      createRetryLambda: false,
+    },
+  };
+  return {
+    writeToS3,
+  };
+}
+
+type QueueAndLambdaSettings = {
+  name: string;
+  entry: string;
+  lambda: {
+    memory: 512 | 1024 | 2048 | 4096;
+    /** Number of messages the lambda pull from SQS at once  */
+    batchSize: number;
+    /** How long can the lambda run for, max is 900 seconds (15 minutes)  */
+    timeout: Duration;
+    /** Partial batch response: https://docs.aws.amazon.com/prescriptive-guidance/latest/lambda-event-filtering-partial-batch-responses-for-sqs/welcome.html */
+    reportBatchItemFailures: boolean;
+  };
+  queue: {
+    alarmMaxAgeOfOldestMessage: Duration;
+    maxMessageCountAlarmThreshold?: number;
+    /** The number of times a message can be unsuccesfully dequeued before being moved to the dead-letter queue. */
+    maxReceiveCount: number;
+    /** How long messages should be invisible for other consumers, based on the lambda timeout */
+    /** We don't care if the message gets reprocessed, so no need to have a huge visibility timeout that makes it harder to move messages to the DLQ */
+    visibilityTimeout: Duration;
+    createRetryLambda: boolean;
+  };
+};
 
 export class IHEGatewayV2LambdasNestedStack extends NestedStack {
   constructor(scope: Construct, id: string, props: IHEGatewayV2LambdasNestedStackProps) {
     super(scope, id, props);
 
     this.terminationProtection = true;
+
+    const { lambda: writeToS3LambdaOutboundPD, queue: writeToS3QueueOutboundPD } =
+      this.setupWriteToS3OutboundPD({
+        lambdaLayers: props.lambdaLayers,
+        vpc: props.vpc,
+        envType: props.envType,
+        sentryDsn: props.sentryDsn,
+        alarmAction: props.alarmAction,
+      });
 
     const iheResponsesBucket = new s3.Bucket(this, "IHEResponsesBucket", {
       bucketName: props.iheResponsesBucketName,
@@ -45,8 +104,6 @@ export class IHEGatewayV2LambdasNestedStack extends NestedStack {
       versioned: true,
     });
 
-    iheResponsesBucket.grantWrite(props.writeToS3Lambda);
-
     const iheParsedResponsesBucket = new s3.Bucket(this, "iheParsedResponsesBucket", {
       bucketName: props.iheParsedResponsesBucketName,
       publicReadAccess: false,
@@ -54,14 +111,15 @@ export class IHEGatewayV2LambdasNestedStack extends NestedStack {
       versioned: true,
     });
 
-    iheParsedResponsesBucket.grantWrite(props.writeToS3Lambda);
+    iheParsedResponsesBucket.grantWrite(writeToS3LambdaOutboundPD);
 
     this.createParsedReponseTables(iheParsedResponsesBucket);
 
     const patientDiscoveryLambda = this.setupIHEGatewayV2PatientDiscoveryLambda(
       props,
       iheResponsesBucket,
-      iheParsedResponsesBucket
+      iheParsedResponsesBucket,
+      writeToS3QueueOutboundPD
     );
     const documentQueryLambda = this.setupIHEGatewayV2DocumentQueryLambda(
       props,
@@ -121,6 +179,62 @@ export class IHEGatewayV2LambdasNestedStack extends NestedStack {
     });
   }
 
+  private setupWriteToS3OutboundPD(ownProps: {
+    lambdaLayers: LambdaLayers;
+    vpc: ec2.IVpc;
+    envType: EnvType;
+    sentryDsn: string | undefined;
+    alarmAction: SnsAction | undefined;
+  }): { lambda: Lambda; queue: Queue } {
+    const { lambdaLayers, vpc, envType, sentryDsn, alarmAction } = ownProps;
+    const {
+      name,
+      entry,
+      lambda: { memory, timeout, batchSize, reportBatchItemFailures },
+      queue: {
+        visibilityTimeout,
+        maxReceiveCount,
+        alarmMaxAgeOfOldestMessage,
+        maxMessageCountAlarmThreshold,
+        createRetryLambda,
+      },
+    } = settings().writeToS3;
+
+    const queue = createQueue({
+      stack: this,
+      name,
+      createDLQ: true,
+      visibilityTimeout,
+      maxReceiveCount,
+      lambdaLayers: [lambdaLayers.shared],
+      envType,
+      alarmSnsAction: alarmAction,
+      alarmMaxAgeOfOldestMessage,
+      maxMessageCountAlarmThreshold,
+      createRetryLambda,
+    });
+
+    const lambda = createLambda({
+      stack: this,
+      name,
+      entry,
+      envType,
+      envVars: {
+        // API_URL set on the api-stack after the OSS API is created
+        ...(sentryDsn ? { SENTRY_DSN: sentryDsn } : {}),
+      },
+      layers: [lambdaLayers.shared],
+      memory: memory,
+      timeout: timeout,
+      vpc,
+      alarmSnsAction: alarmAction,
+    });
+
+    lambda.addEventSource(new SqsEventSource(queue, { batchSize, reportBatchItemFailures }));
+
+    return { lambda, queue };
+  }
+
   private grantSecretsReadAccess(
     lambdaFunction: Lambda,
     secrets: Secrets,
@@ -148,10 +262,10 @@ export class IHEGatewayV2LambdasNestedStack extends NestedStack {
       apiURL: string;
       envType: EnvType;
       sentryDsn: string | undefined;
-      writeToS3Queue: IQueue;
     },
     iheResponsesBucket: s3.Bucket,
-    iheParsedResponsesBucket: s3.Bucket
+    iheParsedResponsesBucket: s3.Bucket,
+    writeToS3Queue: Queue
   ): Lambda {
     const {
       lambdaLayers,
@@ -166,7 +280,6 @@ export class IHEGatewayV2LambdasNestedStack extends NestedStack {
       apiURL,
       envType,
       sentryDsn,
-      writeToS3Queue,
     } = ownProps;
 
     const patientDiscoveryLambda = createLambda({
