+++ packages/lambdas/src/ihe-outbound-patient-discovery-write-to-s3.ts
